Add automatic scaling and saving of cached images in _scaled folder.
The DrawBot Context now knows how to create a scaled image path name:
DrawBotContext.path2ScaledImagePath(path, w, h)
The Image class is saving to a cached file, if the flag is (default) on
and if the scaling is over a 0.8 — 1.2 range in either direction.
Add Element.prepare(view) recursive broadcast to all child elements of
a document and page, to prepare for building. This Image is using this
call to make sure all images exist in the caching size that is needed.
The initial call for the page.prepare(view) is done by the PageView,
jut before any page.build is called.
Add _local to .gitignore
Add more doctests